+++ packages/route-store-dom/src/HistoryHandler.ts
@@ -1,4 +1,4 @@
-import { useStore, useFleurContext } from '@fleur/fleur-react'
+import { useStore, useFleurContext } from '@fleur/react'
 import { LocationListener } from 'history'
 import { useLayoutEffect, useEffect, useCallback } from 'react'
 import { navigateOp } from './operations'
@@ -36,55 +36,49 @@ export const HistoryHandler = () => {
   }, [])
 
   // Apply current store route to location
-  useEffect(
-    () => {
-      const route = progressRoute
+  useEffect(() => {
+    const route = progressRoute
 
-      if (!route) return
+    if (!route) return
 
-      if (route.type === 'POP') {
-        const { state } = history.location
-        if (state) {
-          setTimeout(() => window.scrollTo(state.scrollX, state.scrollY))
-        }
-      } else if (route.type === 'REPLACE') {
-        history.replace(route.url, { fluerHandled: true })
-      } else {
-        history.push(route.url, { fluerHandled: true })
+    if (route.type === 'POP') {
+      const { state } = history.location
+      if (state) {
+        setTimeout(() => window.scrollTo(state.scrollX, state.scrollY))
       }
-    },
-    [progressRoute],
-  )
+    } else if (route.type === 'REPLACE') {
+      history.replace(route.url, { fluerHandled: true })
+    } else {
+      history.push(route.url, { fluerHandled: true })
+    }
+  }, [progressRoute])
 
   // Save scroll position
-  useEffect(
-    () => {
-      let scrollTimerId: number
+  useEffect(() => {
+    let scrollTimerId: number
 
-      const handleScroll = () => {
-        if (scrollTimerId) {
-          clearTimeout(scrollTimerId)
-        }
-
-        scrollTimerId = (setTimeout(() => {
-          if (route) {
-            history.replace(route.url, {
-              scrollX: window.scrollX || window.pageXOffset,
-              scrollY: window.scrollY || window.pageYOffset,
-            })
-          }
-        }, 150) as any) as number
-      }
-
-      window.addEventListener('scroll', handleScroll, { passive: true })
-
-      return () => {
-        window.removeEventListener('scroll', handleScroll)
+    const handleScroll = () => {
+      if (scrollTimerId) {
         clearTimeout(scrollTimerId)
       }
-    },
-    [route],
-  )
+
+      scrollTimerId = (setTimeout(() => {
+        if (route) {
+          history.replace(route.url, {
+            scrollX: window.scrollX || window.pageXOffset,
+            scrollY: window.scrollY || window.pageYOffset,
+          })
+        }
+      }, 150) as any) as number
+    }
+
+    window.addEventListener('scroll', handleScroll, { passive: true })
+
+    return () => {
+      window.removeEventListener('scroll', handleScroll)
+      clearTimeout(scrollTimerId)
+    }
+  }, [route])
 
   return null
 }