+++ packages/fleur/CHANGELOG.md
@@ -1,5 +1,8 @@
 ### 1.2.0
 
+- Add `selector()` and `selectorWithStore()`<br />
+  `selectorWithStore()` is provided for compatibility.<br />
+  Please think to use `selector()` first.
 - Type `StoreGetter` moved from `@fleur/fleur-react`
 
 ### 1.1.2